bin-pipecat-manager/scripts/pipecat/pyproject.toml
[tool.poetry]
name = "pipecat-fastapi-server"
version = "0.1.0"
description = "Quickstart example for building voice AI bots with Pipecat"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"

# === Core ===
fastapi = ">=0.115.0"
uvicorn = {extras = ["standard"], version = ">=0.30.0"}
python-dotenv = ">=1.0.0"
loguru = ">=0.7.0"
pydantic = ">=2.5.0"
asyncio = "*"

# === Pipecat stack ===
pipecat-ai = {extras = ["cartesia", "daily", "deepgram", "local-smart-turn-v3", "openai", "runner", "silero", "webrtc"], version = ">=0.0.82"}
pipecatcloud = ">=0.2.4"

# === Others ===
orjson = ">=3.9.0"

[tool.poetry.group.dev.dependencies]
ruff = "~=0.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["I"]
